unify fonts etc. and move some common functions to ui-utils
