+++ src/clients/artikkelit/artikkelit.css
@@ -1,5 +1,5 @@
 /* CSS file for muuntaja */
-
+@import "/common/stylesheets/fonts.css";
 @import "/common/stylesheets/common.css";
 
 /* ---------------------------------------------------------------------------- */
@@ -10,65 +10,12 @@ div#loading {}
 
 div#tools {}
 
-.tools .ToolBar .Input>label {
-  color: rgb(48, 93, 109);
-  font-family: roboto_monoregular, "Roboto Mono", monospace;
-  font-size: 0.8em;
-  text-align: center;
-}
-
-.tools .ToolBar .Input>input {
-  border: 0;
-  border-bottom: 1pt solid grey;
-  outline: none;
-  font-family: roboto_monoregular, "Roboto Mono", monospace;
-  font-size: 1em;
-  width: 37ch;
-}
-
-.tools .ToolBar .Select>label,
-.header .Select>label {
-  color: rgb(48, 93, 109);
-  font-family: roboto_monoregular, "Roboto Mono", monospace;
-  font-size: 0.8em;
-  text-align: center;
-  text-transform: none;
-}
-
-.tools .ToolBar .Select>select,
-.header .Select>select {
-  border: 0;
-  border-bottom: 1pt solid grey;
-  outline: none;
-  font-family: roboto_monoregular, "Roboto Mono", monospace;
-  font-size: 1em;
-}
-
-.tools #sequence {
-  width: 6ch;
-  text-align: end;
-  margin-left: 1ch;
-}
-
-.tools #logType {
-  text-align: end;
-  margin-right: 1ch;
-}
-
-.tools>.dropdown,
-.tools>.dropdown>button {
-  height: 40px;
-  font-size: 32px;
-}
-
 .tools {
   display: flex;
   flex-direction: row;
   justify-content: center;
   align-items: flex-end;
   width: 100vw;
-  font-size: 32px;
-  height: 40px;
 }
 
 /* ---------------------------------------------------------------------------- */
@@ -116,34 +63,37 @@ div#muuntaja {}
   padding: 8pt;
 }
 
-.col .header {
-  height: 40px;
+.header {
   text-transform: uppercase;
-  font-family: roboto_monoregular, "Roboto Mono", monospace;
+  font-family: var(--font-family-headline);
+  font-size: var(--font-size-headline-normal);
+  font-weight: var(--font-weight-headline-normal);
+  line-height: var(--font-line-height-headline-normal);
   border-bottom: 1px solid lightgrey;
   align-items: center;
   display: flex;
 }
 
-.col .header .title {
-  flex: 4;
-}
-
-.col .hidden {
+.hidden {
   display: none;
 }
 
-.col form {
+form {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  gap: 8px;
   width: 100%;
 }
 
 .header .material-icons {
-  padding: 8px;
-  font-size: 32px;
+  font-size: var(--font-size-icon-normal);
 }
 
 .header label {
-  font-size: 12pt;
+  font-size: var(--font-size-label-normal);
+  font-weight: var(--font-weight-label-normal);
+  line-height: var(--font-line-height-label-normal);
   /*color: grey;*/
 }
 
@@ -156,7 +106,7 @@ div#muuntaja {}
   display: flex;
   align-items: center;
   align-content: center;
-  line-height: 20px;
+  gap: 5px;
 }
 
 .full-width label {
@@ -176,34 +126,30 @@ div#muuntaja {}
 }
 
 .full-width p {
-  margin-top: 2px;
-  margin-bottom: 2px;
+  margin: 0px;
 }
 
 .full-width p:last-of-type {
   flex-grow: 1;
 }
 
-.full-width button {
-  margin-left: 8px;
-}
-
 .capitalize {
   text-transform: capitalize;
 }
 
-form>* {
-  margin-top: 8px;
-}
-
 fieldset {
   display: flex;
-  align-items: center;
-  flex-direction: row;
+  align-items: stretch;
+  flex-direction: column;
   flex-wrap: wrap;
-  gap: 20px;
+  gap: 8px;
   margin-bottom: 20px;
   border-radius: 8px;
+}
+
+fieldset legend {
+  font-size: var(--font-size-title-large);
+  font-weight: var(--font-weight-title-large);
 }
 
 fieldset .panel {
@@ -217,16 +163,16 @@ fieldset .panel>.full-width {
 
 fieldset .panel label {
   display: inline-flex;
-  font-family: "Roboto", "Roboto Mono", monospace;
+  font-family: var(--font-family-label);
   text-align: end;
   margin-right: 10px;
-  line-height: 40px;
+  line-height: var(--font-line-height-label-normal);
   color: rgb(48, 93, 109);
 }
 
 fieldset .panel p {
   display: inline-flex;
-  line-height: 40px;
+  line-height: var(--font-line-height-content-normal);
 }
 
 .long-text {
@@ -238,8 +184,24 @@ fieldset .panel p {
 
 .label-text {
   color: rgb(48, 93, 109);
-  font-family: "Roboto", "Roboto Mono", monospace;
-  font-size: 12pt;
+  font-family: var(--font-family-label);
+  font-size: var(--font-size-label-normal);
+  font-weight: var(--font-weight-label-normal);
+}
+
+.title-text {
+  color: rgb(48, 93, 109);
+  font-family: var(--font-family-title);
+  font-size: var(--font-size-title-normal);
+  font-weight: var(--font-weight-title-normal);
+
+}
+
+.content-text {
+  color: rgb(0, 0, 0);
+  font-family: var(--font-family-content);
+  font-size: var(--font-size-content-normal);
+  font-weight: var(--font-weight-content-normal);
 }
 
 .bold-text {
@@ -247,11 +209,10 @@ fieldset .panel p {
 }
 
 .justify-content-end {
-  justify-content: end;
   display: inline-flex;
+  justify-content: end;
 }
-
-.justify-content-end form {
+.justify-content-end>form {
   width: auto;
 }
