+++ src/clients/artikkelit/artikkelit.js
@@ -1,17 +1,16 @@
-import {setNavBar, startProcess, stopProcess} from "/common/ui-utils.js";
-import {Account, doLogin, logout} from "/common/auth.js"
-import {showTab, resetForms, reload} from "/common/ui-utils.js";
-import {getArtikkeliRecord, getPublicationByISSN, getPublicationByISBN, getPublicationByTitle, getPublicationByMelinda, getOntologyWords} from "../common/rest.js";
+import {setNavBar} from "/common/ui-utils.js";
+import {doLogin} from "/common/auth.js"
+import {showTab} from "/common/ui-utils.js";
+import {getArtikkeliRecord} from "../common/rest.js";
 import {showRecord} from "/common/marc-record-ui.js";
-import {formToJson} from "/common/ui-utils.js";
-import {idbGet, idbSet, idbDel, idbGetStoredValues} from "/artikkelit/indexDB.js"
-import {setOptions} from "/artikkelit/utils.js";
+import {idbGet, idbDel, idbGetStoredValues} from "/artikkelit/indexDB.js"
 import {initAuthors, refreshAuthorsList, refreshAuthorOrganizationList, resetAuthor} from "/artikkelit/interfaces/authors.js";
 import {initAbstracts, refreshAbstractList} from "/artikkelit/interfaces/abstracts.js";
 import {initOntologyWords, refreshOntologyWordList} from "/artikkelit/interfaces/ontologyWords.js";
 import {fillFormOptions, fillDatalistOptions} from "/artikkelit/interfaces/loadData.js";
 import {initArticle, refreshSciencesList, refreshMetodologysList} from "/artikkelit/interfaces/article.js";
 import {initAdditionalFields, refreshNotesList, refreshUDKsList, refreshOtherRatingsList} from "/artikkelit/interfaces/additionalFields.js";
+import {initPublicationSearch} from "./interfaces/publicationSearch.js";
 //import { } from "./interfaces/";
 
 window.initialize = function () {
@@ -25,18 +24,12 @@ window.initialize = function () {
     // username.innerHTML = Account.get()["Name"];
     showTab('artikkelit-lisaa');
     fillFormOptions();
+    initPublicationSearch();
     initArticle();
     initAuthors();
     initAbstracts();
     initOntologyWords();
     initAdditionalFields();
-
-    document.getElementById("julkaisu-haku-title-form").addEventListener("submit", searchPublications);
-    document.getElementById("julkaisu-haku-melinda-form").addEventListener("submit", searchPublications);
-    document.getElementById("julkaisu-haku-isbn-form").addEventListener("submit", searchPublications);
-    document.getElementById("julkaisu-haku-issn-form").addEventListener("submit", searchPublications);
-
-    resetSearchResultSelect();
   }
 }
 
@@ -50,8 +43,7 @@ window.sourceTypeChange = (event) => {
     document.getElementById(`numeron-numero-wrap`).style.display = 'block';
     document.getElementById(`artikkelin-osasto-toistuva-wrap`).style.display = 'block';
     document.getElementById(`artikkelin-arvostelu-tyyppi-wrap`).style.display = 'block';
-    document.getElementById(`lehden-issn-wrap`).style.display = 'block';
-    document.getElementById(`lehden-isbn-wrap`).style.display = 'none';
+    document.getElementById(`lehden-tunniste-label`).innerHTML = 'ISSN:';
   }
 
   if (sourceType === 'book') {
@@ -59,40 +51,7 @@ window.sourceTypeChange = (event) => {
     document.getElementById(`numeron-numero-wrap`).style.display = 'none';
     document.getElementById(`artikkelin-osasto-toistuva-wrap`).style.display = 'none';
     document.getElementById(`artikkelin-arvostelu-tyyppi-wrap`).style.display = 'none';
-    document.getElementById(`lehden-issn-wrap`).style.display = 'none';
-    document.getElementById(`lehden-isbn-wrap`).style.display = 'block';
-  }
-}
-
-window.showAndHideSearchInputs = (event) => {
-  document.getElementById(`julkaisu-haku-title-form`).style.display = 'none';
-  document.getElementById(`julkaisu-haku-issn-form`).style.display = 'none';
-  document.getElementById(`julkaisu-haku-isbn-form`).style.display = 'none';
-  document.getElementById(`julkaisu-haku-melinda-form`).style.display = 'none';
-
-  // console.log(event.target.value);
-
-  document.getElementById(`julkaisu-haku-${event.target.value}-form`).style.display = 'block';
-}
-
-window.ontologySearchResultChange = () => {
-
-}
-
-window.searchResultChange = (event) => {
-  if (event.target.value !== '') {
-    idbGet('artoSources', parseInt(event.target.value)).then(data => {
-      //console.log(data);
-      document.getElementById(`lehden-nimi`).innerHTML = data.title;
-      document.getElementById(`lehden-melindaId`).innerHTML = data.sourceIds.filter(id => (/^\(FI-MELINDA\)\d{9}$/u).test(id));
-      document.getElementById(`lehden-issn`).innerHTML = data.issns;
-      document.getElementById(`lehden-isbn`).innerHTML = data.isbns;
-      document.getElementById(`lehden-julkaisu-tyyppi`).innerHTML = data.recordType;
-      document.getElementById(`lehden-elektroninen-julkaisu`).innerHTML = data.isElectronic ? "Kyllä" : "Ei";
-      document.getElementById(`lehden-vuodet-min`).innerHTML = data.publisherInfo.publisherYears.start;
-      document.getElementById(`lehden-vuodet-max`).innerHTML = data.publisherInfo.publisherYears.end;
-      document.getElementById(`lehden-paikka`).innerHTML = data.publisherInfo.publisherLocation;
-    });
+    document.getElementById(`lehden-tunniste-label`).innerHTML = 'ISBN:';
   }
 }
 
@@ -165,64 +124,6 @@ function collectFormData() {
     }
 
   };
-}
-
-function resetSearchResultSelect(searching) {
-  const select = document.getElementById('julkaisu-haku-tulos-lista');
-  select.innerHTML = '';
-  document.getElementById(`lehden-nimi`).innerHTML = '';
-  document.getElementById(`lehden-melindaId`).innerHTML = '';
-  document.getElementById(`lehden-issn`).innerHTML = '';
-  document.getElementById(`lehden-isbn`).innerHTML = '';
-  document.getElementById(`lehden-julkaisu-tyyppi`).innerHTML = '';
-  document.getElementById(`lehden-elektroninen-julkaisu`).innerHTML = '';
-  document.getElementById(`lehden-vuodet-min`).innerHTML = '';
-  document.getElementById(`lehden-vuodet-max`).innerHTML = '';
-
-  if (searching) {
-    return setOptions(select, [{value: '', text: 'Etsitään...'}], true);
-  }
-
-  setOptions(select, [{value: '', text: 'Ei tuloksia'}], true);
-}
-
-function searchPublications(event) {
-  event.preventDefault();
-
-  resetSearchResultSelect(true);
-  const hakuTyyppi = document.getElementById(`julkaisu-haku-tyyppi`).value;
-  const sourceType = document.querySelector('#kuvailtava-kohde').value;
-  const formJson = formToJson(event);
-
-  if (hakuTyyppi === 'issn') {
-    return getPublicationByISSN(formJson['haku-arvo'], sourceType).then(result => console.log(result));
-  }
-  if (hakuTyyppi === 'isbn') {
-    return getPublicationByISBN(formJson['haku-arvo'], sourceType).then(result => console.log(result));
-  }
-  if (hakuTyyppi === 'title') {
-    return getPublicationByTitle(formJson['haku-arvo'], sourceType).then(result => setRecordsToSearch(result));
-  }
-  if (hakuTyyppi === 'melinda') {
-    return getPublicationByMelinda(formJson['haku-arvo'], sourceType).then(result => console.log(result));
-  }
-
-  throw new Error('Invalid search type!');
-}
-
-function setRecordsToSearch(records) {
-  if (records.length === 0) {
-    return resetSearchResultSelect();
-  }
-
-  const select = document.getElementById('julkaisu-haku-tulos-lista');
-  const data = records.map((record, index) => {
-    const title = record.data.title;
-    idbSet('artoSources', index, record.data);
-    return {value: index, text: title};
-  });
-
-  setOptions(select, data);
 }
 
 window.removeScience = (event, key) => {
