+++ src/clients/artikkelit/interfaces/loadData.js
@@ -1,5 +1,5 @@
-import {setOptions} from "/artikkelit/utils.js"
 import {articleTypesBooks, articleTypesJournal, authorRelators, languages, ontologyTypes, organizations, sciences, searchTypes, sourceTypes} from "/artikkelit/interfaces/constants.js";
+import {setOptions} from "/common/ui-utils.js"
 
 export function fillFormOptions() {
   fillSelectOptions();
